+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - Attack range calculation on actor sheets could cause the sheet to be unopeneable
 - Fly maneuverability bonuses/penalties didn't apply to the fly skill bonus
 - Changing an attack's damage ability multiplier in the attack dialog changed the value for the item in-memory
+- The Wound Threshold optional rule was no longer working
 
 ### Changelog